pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "gService",
  "gDrive"
]

[project]
name = "gDrive_Majoneza"
version = "0.0.3"
authors = [
  { name="Majoneza", email="author@example.com" },
]
description = "Library to manage your Google Drive for Python programming language"
readme = "README.md"
license = {file = 'LICENSE'}
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "google-api-python-client",
  "google-auth-oauthlib",
  "google-auth",
]

[project.urls]
Homepage = "https://github.com/Majoneza/gDrive"

[project.optional-dependencies]
crypto = [
  "pycryptodome",
]
